fix: v1.3 — TP/TCP overflow DoS, SD endpoint options, hardening (#79)
Critical (remote DoS, memory-safety):

- tp.Reassembler.Add trusted an attacker-controlled 28-bit TP Offset field
  with no bound: a crafted segment stored at a large offset followed by a
  final segment with a smaller totalSize reached
  `copy(payload[off:], data)` with off > len(payload), panicking with
  "slice bounds out of range" in the goroutine handling the connection
  (go-SOMEIP-01). A single segment could also claim an offset up to ~4 GiB,
  forcing an unbounded `make([]byte, totalSize)` allocation. Added
  tp.MaxReassembledSize (64 MiB), rejecting any segment whose offset+len
  exceeds it, and made the final reassembly copy defensive against any
  chunk that still falls outside the learned totalSize. AUTOSAR SOME/IP-TP
  defines Offset as a 28-bit field with no spec-mandated trust guarantee;
  RELAY spec §5.4 requires malformed input to yield an error, not a crash.
  Regression tests (TestReassembler_CraftedOffsetDoesNotPanic,
  TestReassembler_OffsetExceedsMaxReassembledSize) reproduce both failure
  modes and are confirmed to panic/misbehave against the pre-fix code.

- tcp.readFrame computed `codec.HeaderSize + payloadLen` in uint32
  arithmetic from an attacker-controlled Length field: Length=0xFFFFFFFF
  wrapped the allocation size to 7 bytes, and the following
  `frame[16:]` slice panicked — a single 16-byte header crashing the
  server. A large non-wrapping Length (e.g. 0x40000000) instead forced a
  ~1 GiB allocation per connection. Length/payload arithmetic now runs in
  int64 with an explicit maxFrameSize (64 MiB) bound, returning
  someip.ErrPayloadTooLarge before any allocation (go-SOMEIP-02).
  Regression tests confirmed to panic/over-allocate against the pre-fix
  code (TestReadFrame_LengthOverflowWraps,
  TestReadFrame_LargeNonWrappingLength).

Medium:

- codec.Decode's `wantTotal := int(8 + length)` mirrored the same
  wrapping-uint32 pattern as tcp.go; not currently reachable (the
  len(b)!=wantTotal check rejects it either way, since no caller presents
  Decode with a real multi-GB buffer) but a landmine if a caller path
  changes. Now computed as `8 + int64(length)` (go-SOMEIP-04).

- sd/udp's decodeSDPayload discarded the SD Options array entirely and
  handleEntry registered OfferService endpoints from the UDP source
  address/port with Protocol hardcoded to UDP, so a client using the
  registry contacted the SD multicast/unicast source port instead of the
  offered service's actual port, and TCP offers were silently registered
  as UDP (go-SOMEIP-03). Added sd.DecodeOptions (bounds-checked walk of
  the Options array) and sd.ResolveIPv4Endpoint (resolves an entry's
  Index1/NumOpts1 + Index2/NumOpts2 option runs against it, ignoring
  out-of-range runs rather than trusting attacker-controlled indices).
  handleEntry now registers the endpoint the entry actually references,
  falling back to the SD source address only when no option resolves.

Low:

- sd.EncodeEntry shifted an unmasked NumOpts1 into byte3, so a value above
  15 (the field's 4-bit width) silently wrapped to an unrelated smaller
  count (17 encoded indistinguishably from 1) instead of failing safe.
  Both NumOpts1 and NumOpts2 now saturate at 15 (go-SOMEIP-05).

- tcp.Server.handleMessage never compared a request's InterfaceVersion
  against the served interface, silently dispatching mismatched-interface
  requests to handlers that assume the configured wire format. Now
  compares and returns RetWrongInterfaceVersion (0x08) per AUTOSAR SOME/IP
  Protocol Specification Table 27 (go-SOMEIP-08); RequestNoReturn messages
  with a mismatched interface are dropped rather than dispatched (no
  response path exists for that message type either way).

- All third-party GitHub Actions were pinned to mutable major-version tags
  (@v5, @v6, ...). Pinned to immutable commit SHAs with a `# vX` comment
  for traceability (go-SOMEIP-07).

Not attempted: go-SOMEIP-06 (E2E Profile 01/05 vs. published AUTOSAR
known-answer test vectors) — needs the actual reference vectors, not a
code change; left as an open advisory item.

+++ codec/codec.go
@@ -97,8 +97,8 @@ func Decode(b []byte) (someip.Message, error) {
 	if length < minLength {
 		return someip.Message{}, fmt.Errorf("%w: length field %d < minimum %d", ErrLengthMismatch, length, minLength)
 	}
-	wantTotal := int(8 + length) // 8 (ServiceID+MethodID+Length) + length
-	if len(b) != wantTotal {
+	wantTotal := 8 + int64(length) // 8 (ServiceID+MethodID+Length) + length, widened to avoid uint32 overflow
+	if int64(len(b)) != wantTotal {
 		return someip.Message{}, fmt.Errorf("%w: frame is %d bytes, length field implies %d", ErrLengthMismatch, len(b), wantTotal)
 	}
